config file reading - currently needs to be CJS
